Make Print action print to the proper io::Write object
The Print action prints data straight to stdout, which is not what users
expect from a context where custom stdout and stderr io::Write objects
can be supplied.
This change fixes the problem. It piggy-backs on the Exit variant of the
ParseResult enum and adds an optional String argument to it. This
argument, if present, will be printed to the correct io::Write object
that the user intended to represent stdout, just before exiting the
program.